Add GetModuleBase API
This API is not exposed at the module level, rather the callback provided to ReadFromProcess is invoked with an object that has GetModuleBase as a key.

GetModuleBase takes a name of a module as an argument, and then searches the process' loaded modules for it and returns its base memory offset.